+++ R/project.R
@@ -13,6 +13,8 @@ NULL
 #'
 #' @param x A [`BiodiversityScenario`] object with set predictors. Note that some
 #' constrains such as \code{MigClim} can still simulate future change without projections.
+#' Alternatively a [`DistributionModel`] object can be supplied if other scenario
+#' functions are not needed. In this case provide a \code{env} parameter value.
 #' @param date_interpolation A [`character`] on whether dates should be interpolated.
 #' Options include \code{"none"} (Default), \code{"annual"}, \code{"monthly"}, \code{"daily"}.
 #' @param stabilize A [`logical`] value indicating whether the suitability projection
@@ -20,6 +22,8 @@ NULL
 #' @param stabilize_method [`character`] stating the stabilization method to be
 #' applied. Currently supported is \code{`loess`}.
 #' @param layer A [`character`] specifying the layer to be projected (Default: \code{"mean"}).
+#' @param env An optional [`SpatRaster`] or [`data.frame`] object for prediction.
+#' Ignored unless a [`DistributionModel`] object is supplied (Default: \code{NULL}).
 #' @param verbose Setting this [`logical`] value to \code{TRUE} prints out further
 #' information during the model fitting (Default: \code{FALSE}).
 #' @param ... passed on parameters.
@@ -97,7 +101,7 @@ methods::setMethod(
   "project",
   methods::signature(x = "BiodiversityScenario"),
   function(x, date_interpolation = "none", stabilize = FALSE, stabilize_method = "loess",
-           layer = "mean", verbose = getOption('ibis.setupmessages', default = TRUE), ...){
+           layer = "mean", env = NULL, verbose = getOption('ibis.setupmessages', default = TRUE), ...){
     # MJ: Workaround to ensure project generic does not conflict with terra::project
     mod <- x
     # date_interpolation = "none"; stabilize = FALSE; stabilize_method = "loess"; layer="mean"
@@ -739,6 +743,74 @@ methods::setMethod(
     out <- mod$clone(deep = TRUE)
     out$scenarios <- proj
     out$scenarios_migclim <- mc
+    return(out)
+  }
+)
+
+#' @rdname project
+#' @export
+project.DistributionModel <- function(x,...) project(x,...)
+
+#' @rdname project
+#' @export
+methods::setMethod(
+  "project",
+  methods::signature(x = "DistributionModel"),
+  function(x, env, layer = "mean", verbose = getOption('ibis.setupmessages', default = TRUE), ...){
+    assertthat::assert_that(
+      is.Raster(env) || is.data.frame(env),
+      is.character(layer)
+    )
+    # Get coefficients and model
+    # co <- x$get_coefficients()[,1]
+    co <- x$model$biodiversity[[1]]$predictors_names
+    model <- x$model
+    settings <- x$settings
+
+    # Further checks
+    assertthat::assert_that(
+      length(co)>0,
+      is.list(model)
+    )
+    # If names are to be sanitized, cleanl
+    if(settings$get("ibis.cleannames")){
+      nn <- sanitize_names(names(env))
+      names(env) <- nn
+    }
+
+    # Check that all predictor names are present
+    assertthat::assert_that(
+      all(co %in% names(env)),
+      msg = "Not all coefficients are found in the fitted model..."
+    )
+
+    # Make a template
+    if(is.Raster(env)) {
+      template <- emptyraster(env)
+    } else {
+      assertthat::assert_that(
+        hasName(env, "x") && hasName(env, "y"),
+        msg = "Coordinates as x and y need to be supplied!"
+      )
+      # Create template
+      template <- try({
+        terra::rast(env[,c("x", "y")],
+                    crs = terra::crs(model$background),
+                    type = "xyz") |>
+          emptyraster()
+      },silent = TRUE)
+    }
+
+    # If raster convert to data.frame for further predictions
+    if(is.Raster(env)) env <- as.data.frame(env, xy = TRUE)
+
+    # --- #
+    # Now predict
+    out <- x$project(newdata = env, layer = layer)
+    names(out) <- paste0("suitability", "_", layer)
+    if(is.na(terra::crs(out))) terra::crs(out) <- terra::crs( model$background )
+    # --- #
+
     return(out)
   }
 )